tree 구조 메모리 leak 문제 해결
기존 tree 만드는 로직을 check 와 tree 만드는 부분으로 분리함.
pipe 시에 불필요한 메모리 할당 코드 삭제
